requirements.txt
transformers
torch
openai-whisper
pyaudio
SpeechRecognition
TTS
sounddevice
tensorboard
git+https://github.com/MiscellaneousStuff/whisper.git
